Move prisoner processing into separate class with managed transaction
Ready for when we need to publish event per prisoner after processing,
we need to ensure the prisoners transaction has fully committed before
raising the event.
