+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitallocationapi/service/PrisonerRetryService.kt
@@ -4,18 +4,19 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.context.annotation.Lazy
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-import uk.gov.justice.digital.hmpps.visitallocationapi.model.entity.PrisonerDetails
+import uk.gov.justice.digital.hmpps.visitallocationapi.clients.IncentivesClient
+import uk.gov.justice.digital.hmpps.visitallocationapi.clients.PrisonerSearchClient
+import uk.gov.justice.digital.hmpps.visitallocationapi.dto.incentives.PrisonIncentiveAmountsDto
+import uk.gov.justice.digital.hmpps.visitallocationapi.service.AllocationService.Companion.LOG
 import uk.gov.justice.digital.hmpps.visitallocationapi.service.sqs.VisitAllocationPrisonerRetrySqsService
-import java.time.LocalDate
 
 @Service
 class PrisonerRetryService(
   private val visitAllocationPrisonerRetrySqsService: VisitAllocationPrisonerRetrySqsService,
+  private val incentivesClient: IncentivesClient,
+  private val prisonerSearchClient: PrisonerSearchClient,
   @Lazy
-  private val allocationService: AllocationService,
-  @Lazy
-  private val prisonerDetailsService: PrisonerDetailsService,
+  private val processPrisonerService: ProcessPrisonerService,
 ) {
   companion object {
     val log: Logger = LoggerFactory.getLogger(this::class.java)
@@ -31,15 +32,22 @@ class PrisonerRetryService(
     }
   }
 
-  @Transactional
-  fun handlePrisonerRetry(prisonerId: String) {
+  fun handlePrisonerRetry(jobReference: String, prisonerId: String) {
     log.info("handle prisoner - $prisonerId on retry queue")
+    val prisoner = prisonerSearchClient.getPrisonerById(prisonerId)
+    val allIncentiveLevels = getIncentiveLevelsForPrison(prisonId = prisoner.prisonId)
+    processPrisonerService.processPrisoner(prisonerId, jobReference, allIncentiveLevels, fromRetryQueue = true)
+  }
 
-    val dpsPrisonerDetails: PrisonerDetails = prisonerDetailsService.getPrisonerDetails(prisonerId)
-      ?: prisonerDetailsService.createPrisonerDetails(prisonerId, LocalDate.now().minusDays(14), null)
+  private fun getIncentiveLevelsForPrison(prisonId: String): List<PrisonIncentiveAmountsDto> {
+    val incentiveLevelsForPrison = try {
+      incentivesClient.getPrisonIncentiveLevels(prisonId)
+    } catch (e: Exception) {
+      val failureMessage = "failed to get incentive levels by prisonId - $prisonId in retry queue consumer"
+      LOG.error(failureMessage, e)
+      throw e
+    }
 
-    allocationService.processPrisonerAllocation(dpsPrisonerDetails)
-
-    prisonerDetailsService.updatePrisonerDetails(dpsPrisonerDetails)
+    return incentiveLevelsForPrison
   }
 }
